chore: 컴포넌트 기능별 세부 폴더로 한 번 더 분류
