+++ src/components/button/MultiSelectButton/index.ts
@@ -1,0 +1,1 @@
+export { default as MultiSelectButton } from "./MultiSelectButton";
