+++ src/components/button/SubMenuButton/SubMenuButton.styles.ts
@@ -2,7 +2,7 @@
 import styled from "styled-components";
 import theme from "../../../style/theme";
 
-export const StyledSubMenu = styled.div`
+export const StyledSubMenuButton = styled.div`
   display: flex;
   align-items: center;
   cursor: pointer;
