+++ src/components/button/SubMenuButton/index.ts
@@ -1,0 +1,1 @@
+export { default as SubMenuButton } from "./SubMenuButton";
