--- src/components/global/ButtonGrid/index.ts
@@ -1,1 +1,0 @@
-export { default as ButtonGrid } from "./ButtonGrid";
