--- src/components/global/SubMenu/index.ts
@@ -1,1 +1,0 @@
-export { default as SubMenu } from "./SubMenu";
