+++ src/components/modal/CustomModal/CustomModal.styles.ts
