+++ src/components/progress/ProgressBar/ProgressBar.styles.ts
