+++ src/components/progress/ProgressBar/index.ts
