+++ src/components/progress/ProgressBlock/ProgressBlock.styles.ts
