+++ src/components/progress/ProgressBlock/ProgressBlock.tsx
