+++ src/components/progress/StepProgress/StepProgress.styles.ts
