+++ src/components/progress/StepProgress/StepProgress.tsx
